refactored seedShows.js to require the entire models folder, started seedUsers.js
