+++ db/seedUsers.js
@@ -1,0 +1,2 @@
+const User = require('../models');
+require('../config/db.connection.js');
