+++ models/index.js
@@ -1,3 +1,5 @@
+require('../config/db.connection');
+
 module.exports = {
     Episode: require('./Episode'),
     Show: require('./Show'),
